Dictinary_Exercise_biggest
We want to write some simple procedures that work on dictionaries to return information. 
This time, write a procedure, called biggest, which returns the key corresponding to the entry with the largest number of values associated with it. If there is more than one such entry, return any one of the matching keys.
Example usage: 
>>> biggest(animals)
'd'
If there are no values in the dictionary, biggest should return None.